backend/pyproject.toml
[project]
name = "hidden-gem-backend"
version = "0.1.0"
description = "Backend for Hidden Gem game collection website"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic[email]>=2.5.0",
    "python-multipart>=0.0.6",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "sqlalchemy>=2.0.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
]

